Overhaul title conceptualisation; fix bugs; add copydirs
Titles are conceptualized as internal (what is shown on the page
itself) and external (what is shown in links, etc.) rather than
brief or full (usage is more clear a concept than brevity).

Some bugs are fixed (e.g. long links used to become "<ahref...>"
instead of "<a href>", which has now been fixed)

added the concept of copydirs: directories from which files are
simply copied into the website root.